backend/requirements.txt
accelerate==0.21.0
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.7.2
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
backcall==0.2.0
beautifulsoup4==4.12.2
bitsandbytes==0.41.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
cmake==3.27.2
constantly==15.1.0
cryptography==41.0.3
dataclasses-json==0.5.14
decorator==5.1.1
defusedxml==0.7.1
diskcache==5.6.1
Django==4.2.4
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework==3.14.0
exceptiongroup==1.1.3
executing==1.2.0
filelock==3.12.2
frozenlist==1.4.0
fsspec==2023.6.0
google-api-core==2.11.1
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.60.0
greenlet==2.0.2
h11==0.14.0
httplib2==0.22.0
huggingface-hub==0.16.4
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.8.0
incremental==22.10.0
ipython==8.14.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
langchain==0.0.266
langsmith==0.0.24
lit==16.0.6
llama-cpp-python==0.1.77
llama-index==0.8.5.post1
Markdown==3.4.4
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
mysqlclient==2.2.0
nest-asyncio==1.5.7
networkx==3.1
nltk==3.8.1
numexpr==2.8.5
numpy==1.25.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
openai==0.27.8
openapi-schema-pydantic
packaging==23.1
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
prompt-toolkit==3.0.39
protobuf==4.24.0
psutil==5.9.5
psycopg2==2.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic_core==2.6.0
Pygments==2.16.1
PyJWT==2.8.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.3.2
scikit-learn==1.3.0
scipy==1.11.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
soupsieve==2.4.1
SQLAlchemy==2.0.20
sqlparse==0.4.4
sse-starlette==1.6.5
stack-data==0.6.2
starlette==0.27.0
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.4.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
traitlets==5.9.0
transformers==4.31.0
triton==2.0.0
Twisted==22.10.0
txaio==23.1.1
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.23.2
wcwidth==0.2.6
websockets==11.0.3
yarl==1.9.2
zipp==3.16.2
zope.interface==6.0
